The io_driver template as a virtual async_io_stream interface.
It is interesting how much simpler the code is using a virtual
function interface. The basic problem with this design is object
ownership. The class needs to reference back to itself when
calling async_read() and async_write(). Since these objects
always are allocated dynamically, we would like to store them in
shared pointers -- but how does the object know its own shared
pointer?

Boost's shared pointers have a way to cast 'this' to a
shared_ptr<T> pointing to this, but does a tr1::shared_ptr
support that too?

The template code, on the other hand, uses shared_ptr<context>
unconditionally. It solves the problem once and for all, but it
enforces a reference-counting approach. The abstract API could,
in theory, perform some sort of self-delete and rely on the state
machine to make sure no object leaks. Whether that is really
possible in the presence of exceptions is a different question.